pyproject.toml
[tool.poetry]
name = "teste-backend-amicci"
version = "0.1.0"
description = ""
authors = ["João Pedro Prado <joaopedro0128@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
gunicorn = "^22.0.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2 = "^2.9.9"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
dj-database-url = "^2.1.0"
whitenoise = "^6.6.0"
black = "^24.4.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
